multicall: Move providers feature implementation in another file
